old_version/flask_app/requirements.txt
flask==2.3.2
psycopg2-binary==2.9.9
werkzeug==2.3.7
requests==2.31.0

